+++ src/Buttons.h
@@ -1,0 +1,18 @@
+#ifndef BUTTONS_H
+#define BUTTONS_H
+
+void Buttons_Setup();
+
+void Buttons_Loop();
+
+bool Get_Button1_ShortPress();
+bool Get_Button1_LongPress();
+
+bool Get_Button2_ShortPress();
+bool Get_Button2_LongPress();
+
+bool Get_Button3_ShortPress();
+bool Get_Button3_LongPress();
+
+
+#endif
